Adding SHA256 Hashing Support For OFSC19C And Above
